refactor(dto): renombrar clases de DTO para seguir una convención consistente
Se añadieron sufijos `DTO` a las clases de solicitud y respuesta para mejorar la claridad y mantener un estándar uniforme en su nomenclatura.
