+++ src/main/java/com/recolectaedu/dto/request/BibliotecaRequestDTO.java
@@ -3,6 +3,6 @@ package com.recolectaedu.dto.request;
 import lombok.Data;
 
 @Data
-public class BibliotecaRequest {
+public class BibliotecaRequestDTO {
     private Integer id_recurso;
 }
