+++ src/main/java/com/recolectaedu/dto/request/ComentarioRequestDTO.java
@@ -3,8 +3,9 @@ package com.recolectaedu.dto.request;
 import lombok.Data;
 
 @Data
-public class ForoRequest {
-    private String titulo;
+public class ComentarioRequestDTO {
     private String contenido;
     private Integer id_usuario;
+    private Integer id_foro;
+    private Integer id_comentario_padre; // Opcional
 }
