+++ src/main/java/com/recolectaedu/dto/request/ForoRequestDTO.java
@@ -3,9 +3,8 @@ package com.recolectaedu.dto.request;
 import lombok.Data;
 
 @Data
-public class ComentarioRequest {
+public class ForoRequestDTO {
+    private String titulo;
     private String contenido;
     private Integer id_usuario;
-    private Integer id_foro;
-    private Integer id_comentario_padre; // Opcional
 }
