+++ src/main/java/com/recolectaedu/dto/request/RecursoRequestDTO.java
@@ -1,19 +1,18 @@
-package com.recolectaedu.dto.response;
+package com.recolectaedu.dto.request;
 
 import com.recolectaedu.model.enums.FormatoRecurso;
 import com.recolectaedu.model.enums.Tipo_recurso;
 import lombok.Data;
-import java.time.LocalDateTime;
 
 @Data
-public class RecursoResponse {
-    private Integer id_recurso;
+public class RecursoRequestDTO {
     private String titulo;
     private String descripcion;
     private String contenido;
     private FormatoRecurso formato;
     private Tipo_recurso tipo;
-    private LocalDateTime creado_el;
+    private Integer año;
+    private Integer periodo;
     private Integer id_usuario;
     private Integer id_curso;
 }
