+++ src/main/java/com/recolectaedu/dto/response/BibliotecaResponseDTO.java
@@ -3,7 +3,7 @@ package com.recolectaedu.dto.response;
 import lombok.Data;
 
 @Data
-public class BibliotecaResponse {
+public class BibliotecaResponseDTO {
     private Integer id_biblioteca;
     private String nombre;
     private Integer id_usuario;
