+++ src/main/java/com/recolectaedu/dto/response/ComentarioResponseDTO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 import java.time.LocalDateTime;
 
 @Data
-public class ForoResponse {
-    private Integer id_foro;
-    private String titulo;
+public class ComentarioResponseDTO {
+    private Integer id_comentario;
+    private String contenido;
     private LocalDateTime creado_el;
     private Integer id_usuario;
 }
