+++ src/main/java/com/recolectaedu/dto/response/CursoResponseDTO.java
@@ -3,10 +3,9 @@ package com.recolectaedu.dto.response;
 import lombok.Data;
 
 @Data
-public class PerfilResponse {
-    private String nombre;
-    private String apellidos;
-    private Short ciclo;
-    private String carrera;
+public class CursoResponseDTO {
+    private Integer id_curso;
     private String universidad;
+    private String nombre;
+    private String carrera;
 }
