+++ src/main/java/com/recolectaedu/dto/response/ForoResponseDTO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 import java.time.LocalDateTime;
 
 @Data
-public class ComentarioResponse {
-    private Integer id_comentario;
-    private String contenido;
+public class ForoResponseDTO {
+    private Integer id_foro;
+    private String titulo;
     private LocalDateTime creado_el;
     private Integer id_usuario;
 }
