+++ src/main/java/com/recolectaedu/dto/response/PerfilResponseDTO.java
@@ -3,9 +3,10 @@ package com.recolectaedu.dto.response;
 import lombok.Data;
 
 @Data
-public class CursoResponse {
-    private Integer id_curso;
-    private String universidad;
+public class PerfilResponseDTO {
     private String nombre;
+    private String apellidos;
+    private Short ciclo;
     private String carrera;
+    private String universidad;
 }
