+++ src/main/java/com/recolectaedu/dto/response/ResenaResponseDTO.java
@@ -1,11 +1,11 @@
-package com.recolectaedu.dto.request;
+package com.recolectaedu.dto.response;
 
 import lombok.Data;
 
 @Data
-public class ResenaRequest {
+public class ResenaResponseDTO {
+    private Integer id_reseña;
     private String titulo;
-    private String contenido;
     private Boolean es_positivo;
     private Integer id_usuario;
     private Integer id_recurso;
